restructure requestion function
