+++ misc.go
@@ -2,37 +2,18 @@ package btcpay
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"net/http"
 )
 
-type LanguageCodesRespose struct {
-	Code            string `json:"code"`
-	CurrentLanguage string `json:"currentLanguage"`
-}
-
-// Bug
 func (c *Client) GetLanguageCodes(ctx context.Context) ([]*LanguageCodesRespose, int, error) {
 	endpoint := fmt.Sprintf("%s/misc/lang", c.URL)
-	req, err := http.NewRequestWithContext(ctx, "GET", endpoint, nil)
-	if err != nil {
-		return nil, 0, err
-	}
-	bytes, statusCode, err := c.doRequest(req)
+	var dataRes []*LanguageCodesRespose
+	statusCode, err := c.doRequest(ctx, endpoint, "GET", &dataRes, nil)
 	if err != nil {
 		return nil, statusCode, err
 	}
-	var dataRes []*LanguageCodesRespose
-	err = json.Unmarshal(bytes, &dataRes)
-	if err != nil {
-		return nil, 0, err
-	}
 	return dataRes, statusCode, nil
-}
-
-type InvoiceCheckoutPage struct {
-	Page []byte
 }
 
 // View the checkout page of an invoice
@@ -48,7 +29,7 @@ func (c *Client) GetInvoiceCheckoutPage(ctx context.Context, invoiceID *InvoiceI
 	if err != nil {
 		return nil, 0, err
 	}
-	bytes, statusCode, err := c.doRequest(req)
+	bytes, statusCode, err := c.doSimpleRequest(req)
 	if err != nil {
 		return nil, statusCode, err
 	}
@@ -65,7 +46,7 @@ func (i *Invoice) GetInvoiceCheckoutPage(ctx context.Context) (*InvoiceCheckoutP
 	if err != nil {
 		return nil, 0, err
 	}
-	bytes, statusCode, err := i.Client.doRequest(req)
+	bytes, statusCode, err := i.Client.doSimpleRequest(req)
 	if err != nil {
 		return nil, statusCode, err
 	}
