+++ notifications.go
@@ -1,47 +1,22 @@
 package btcpay
 
 import (
-	"bytes"
 	"context"
 	"encoding/json"
 	"fmt"
-	"net/http"
 )
-
-type Notification struct {
-	ID     NotificationID
-	Client *Client
-}
-
-type NotificationID string
-
-type NotificationResponse struct {
-	ID          NotificationID `json:"id"`
-	Body        string         `json:"body"`
-	Link        string         `json:"link"`
-	CreatedTime int64          `json:"createdTime"`
-	Seen        bool           `json:"seen"`
-}
 
 func (c *Client) GetNotifications(ctx context.Context, seen ...bool) ([]*NotificationResponse, int, error) {
 	var endpoint string
+	var dataRes []*NotificationResponse
 	if len(seen) > 0 {
 		endpoint = fmt.Sprintf("%s/api/v1/users/me/notifications?%t", c.URL, seen[0])
 	} else {
 		endpoint = fmt.Sprintf("%s/api/v1/users/me/notifications", c.URL)
 	}
-	req, err := http.NewRequestWithContext(ctx, "GET", endpoint, nil)
-	if err != nil {
-		return nil, 0, err
-	}
-	bytes, statusCode, err := c.doRequest(req)
+	statusCode, err := c.doRequest(ctx, endpoint, "GET", &dataRes, nil)
 	if err != nil {
 		return nil, statusCode, err
-	}
-	var dataRes []*NotificationResponse
-	err = json.Unmarshal(bytes, &dataRes)
-	if err != nil {
-		return nil, 0, err
 	}
 	return dataRes, statusCode, nil
 }
@@ -49,42 +24,22 @@ func (c *Client) GetNotifications(ctx context.Context, seen ...bool) ([]*Notific
 // View information about the specified notification
 func (c *Client) GetNotification(ctx context.Context, notificationID *NotificationID) (*NotificationResponse, int, error) {
 	endpoint := fmt.Sprintf("%s/api/v1/users/me/notifications/%s", c.URL, *notificationID)
-	req, err := http.NewRequestWithContext(ctx, "GET", endpoint, nil)
-	if err != nil {
-		return nil, 0, err
-	}
-	bytes, statusCode, err := c.doRequest(req)
+	var dataRes NotificationResponse
+	statusCode, err := c.doRequest(ctx, endpoint, "GET", &dataRes, nil)
 	if err != nil {
 		return nil, statusCode, err
-	}
-	var dataRes NotificationResponse
-	err = json.Unmarshal(bytes, &dataRes)
-	if err != nil {
-		return nil, 0, err
 	}
 	return &dataRes, statusCode, nil
 }
 
 func (n *Notification) GetNotification(ctx context.Context) (*NotificationResponse, int, error) {
 	endpoint := fmt.Sprintf("%s/api/v1/users/me/notifications/%s", n.Client.URL, n.ID)
-	req, err := http.NewRequestWithContext(ctx, "GET", endpoint, nil)
-	if err != nil {
-		return nil, 0, err
-	}
-	bytes, statusCode, err := n.Client.doRequest(req)
+	var dataRes NotificationResponse
+	statusCode, err := n.Client.doRequest(ctx, endpoint, "GET", &dataRes, nil)
 	if err != nil {
 		return nil, statusCode, err
 	}
-	var dataRes NotificationResponse
-	err = json.Unmarshal(bytes, &dataRes)
-	if err != nil {
-		return nil, 0, err
-	}
 	return &dataRes, statusCode, nil
-}
-
-type UpdateNotification struct {
-	Seen bool `json:"seen"`
 }
 
 // Updates the notification
@@ -100,18 +55,10 @@ func (c *Client) UpdateNotification(ctx context.Context, notificationID *Notific
 	if err != nil {
 		return nil, 0, err
 	}
-	req, err := http.NewRequestWithContext(ctx, "PUT", endpoint, bytes.NewBuffer(dataReq))
-	if err != nil {
-		return nil, 0, err
-	}
-	bytes, statusCode, err := c.doRequest(req)
+	var dataRes NotificationResponse
+	statusCode, err := c.doRequest(ctx, endpoint, "PUT", &dataRes, dataReq)
 	if err != nil {
 		return nil, statusCode, err
-	}
-	var dataRes NotificationResponse
-	err = json.Unmarshal(bytes, &dataRes)
-	if err != nil {
-		return nil, 0, err
 	}
 	return &dataRes, statusCode, nil
 }
@@ -125,22 +72,13 @@ func (n *Notification) UpdateNotification(ctx context.Context, updateNotificatio
 	} else {
 		dataReq = []byte(`{"seen":null}`)
 	}
-	fmt.Println(string(dataReq))
 	if err != nil {
 		return nil, 0, err
-	}
-	req, err := http.NewRequestWithContext(ctx, "PUT", endpoint, bytes.NewBuffer(dataReq))
-	if err != nil {
-		return nil, 0, err
-	}
-	bytes, statusCode, err := n.Client.doRequest(req)
-	if err != nil {
-		return nil, statusCode, err
 	}
 	var dataRes NotificationResponse
-	err = json.Unmarshal(bytes, &dataRes)
+	statusCode, err := n.Client.doRequest(ctx, endpoint, "PUT", &dataRes, dataReq)
 	if err != nil {
-		return nil, 0, err
+		return nil, statusCode, err
 	}
 	return &dataRes, statusCode, nil
 }
@@ -148,11 +86,7 @@ func (n *Notification) UpdateNotification(ctx context.Context, updateNotificatio
 // Removes the specified notification.
 func (c *Client) RemoveNotification(ctx context.Context, notificationID *NotificationID) (int, error) {
 	endpoint := fmt.Sprintf("%s/api/v1/users/me/notifications/%s", c.URL, *notificationID)
-	req, err := http.NewRequestWithContext(ctx, "DELETE", endpoint, nil)
-	if err != nil {
-		return 0, err
-	}
-	_, statusCode, err := c.doRequest(req)
+	statusCode, err := c.doRequest(ctx, endpoint, "DELETE", nil, nil)
 	if err != nil {
 		return statusCode, err
 	}
@@ -161,11 +95,7 @@ func (c *Client) RemoveNotification(ctx context.Context, notificationID *Notific
 
 func (n *Notification) RemoveNotification(ctx context.Context) (int, error) {
 	endpoint := fmt.Sprintf("%s/api/v1/users/me/notifications/%s", n.Client.URL, n.ID)
-	req, err := http.NewRequestWithContext(ctx, "DELETE", endpoint, nil)
-	if err != nil {
-		return 0, err
-	}
-	_, statusCode, err := n.Client.doRequest(req)
+	statusCode, err := n.Client.doRequest(ctx, endpoint, "DELETE", nil, nil)
 	if err != nil {
 		return statusCode, err
 	}
