+++ payment_request.go
@@ -1,91 +1,27 @@
 package btcpay
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
 	"fmt"
-	"net/http"
 )
-
-type PaymentRequestID string
-
-type PaymentRequest struct {
-	Store  *Store
-	Client *Client
-	ID     PaymentRequestID
-}
-
-// Enums PaymentRequestStatus
-type BTCPayPaymentRequestStatus string
-
-type PaymentRequestStatus struct {
-	Pending   BTCPayInvoiceStatus
-	Completed BTCPayInvoiceStatus
-	Expired   BTCPayInvoiceStatus
-}
-
-func GetPaymentRequestStatus() *PaymentRequestStatus {
-	return &PaymentRequestStatus{
-		Pending:   "Pending",
-		Completed: "Completed",
-		Expired:   "Expired",
-	}
-}
-
-type PaymentRequestRequest struct {
-	Amount                    float64 `json:"amount"`
-	Title                     string  `json:"title"`
-	Currency                  string  `json:"currency"`
-	Email                     string  `json:"email,omitempty"`
-	Description               string  `json:"description,omitempty"`
-	ExpiryDate                int64   `json:"expiryDate,omitempty"`
-	EmbeddedCSS               string  `json:"embeddedCSS,omitempty"`
-	CustomCSSLink             string  `json:"customCSSLink,omitempty"`
-	AllowCustomPaymentAmounts bool    `json:"allowCustomPaymentAmounts,omitempty"`
-}
-
-type PaymentRequestResponse struct {
-	ID       PaymentRequestID           `json:"id"`
-	Status   BTCPayPaymentRequestStatus `json:"status"`
-	Created  string                     `json:"created"`
-	Archived bool                       `json:"archived"`
-	PaymentRequestRequest
-}
 
 // View information about the existing payment requests
 func (c *Client) GetPaymentRequests(ctx context.Context, storeID *StoreID) ([]*PaymentRequestResponse, int, error) {
 	endpoint := fmt.Sprintf("%s/api/v1/stores/%s/payment-requests", c.URL, *storeID)
-	req, err := http.NewRequestWithContext(ctx, "GET", endpoint, nil)
-	if err != nil {
-		return nil, 0, err
-	}
-	bytes, statusCode, err := c.doRequest(req)
+	var dataRes []*PaymentRequestResponse
+	statusCode, err := c.doRequest(ctx, endpoint, "GET", &dataRes, nil)
 	if err != nil {
 		return nil, statusCode, err
-	}
-	var dataRes []*PaymentRequestResponse
-	err = json.Unmarshal(bytes, &dataRes)
-	if err != nil {
-		return nil, 0, err
 	}
 	return dataRes, statusCode, nil
 }
 
 func (s *Store) GetPaymentRequests(ctx context.Context) ([]*PaymentRequestResponse, int, error) {
 	endpoint := fmt.Sprintf("%s/api/v1/stores/%s/payment-requests", s.Client.URL, s.ID)
-	req, err := http.NewRequestWithContext(ctx, "GET", endpoint, nil)
-	if err != nil {
-		return nil, 0, err
-	}
-	bytes, statusCode, err := s.Client.doRequest(req)
+	var dataRes []*PaymentRequestResponse
+	statusCode, err := s.Client.doRequest(ctx, endpoint, "GET", &dataRes, nil)
 	if err != nil {
 		return nil, statusCode, err
-	}
-	var dataRes []*PaymentRequestResponse
-	err = json.Unmarshal(bytes, &dataRes)
-	if err != nil {
-		return nil, 0, err
 	}
 	return dataRes, statusCode, nil
 }
@@ -93,99 +29,51 @@ func (s *Store) GetPaymentRequests(ctx context.Context) ([]*PaymentRequestRespon
 // Create a new payment request
 func (c *Client) CreatePaymentRequest(ctx context.Context, storeID *StoreID, paymentRequestRequest *PaymentRequestRequest) (*PaymentRequestResponse, int, error) {
 	endpoint := fmt.Sprintf("%s/api/v1/stores/%s/payment-requests", c.URL, *storeID)
-	dataReq, err := json.Marshal(paymentRequestRequest)
-	if err != nil {
-		return nil, 0, err
-	}
-	req, err := http.NewRequestWithContext(ctx, "POST", endpoint, bytes.NewBuffer(dataReq))
-	if err != nil {
-		return nil, 0, err
-	}
-	bytes, statusCode, err := c.doRequest(req)
+	var dataRes PaymentRequestResponse
+	statusCode, err := c.doRequest(ctx, endpoint, "POST", &dataRes, paymentRequestRequest)
 	if err != nil {
 		return nil, statusCode, err
 	}
-	var dataRes *PaymentRequestResponse
-	err = json.Unmarshal(bytes, &dataRes)
-	if err != nil {
-		return nil, 0, err
-	}
-	return dataRes, statusCode, nil
+	return &dataRes, statusCode, nil
 }
 
 func (s *Store) CreatePaymentRequest(ctx context.Context, paymentRequestRequest *PaymentRequestRequest) (*PaymentRequestResponse, int, error) {
 	endpoint := fmt.Sprintf("%s/api/v1/stores/%s/payment-requests", s.Client.URL, s.ID)
-	dataReq, err := json.Marshal(paymentRequestRequest)
-	if err != nil {
-		return nil, 0, err
-	}
-	req, err := http.NewRequestWithContext(ctx, "POST", endpoint, bytes.NewBuffer(dataReq))
-	if err != nil {
-		return nil, 0, err
-	}
-	bytes, statusCode, err := s.Client.doRequest(req)
+	var dataRes PaymentRequestResponse
+	statusCode, err := s.Client.doRequest(ctx, endpoint, "POST", &dataRes, paymentRequestRequest)
 	if err != nil {
 		return nil, statusCode, err
 	}
-	var dataRes *PaymentRequestResponse
-	err = json.Unmarshal(bytes, &dataRes)
-	if err != nil {
-		return nil, 0, err
-	}
-	return dataRes, statusCode, nil
+	return &dataRes, statusCode, nil
 }
 
 // View information about the specified payment request
 func (c *Client) GetPaymentRequest(ctx context.Context, storeID *StoreID, paymentRequestID *PaymentRequestID) (*PaymentRequestResponse, int, error) {
 	endpoint := fmt.Sprintf("%s/api/v1/stores/%s/payment-requests/%s", c.URL, *storeID, *paymentRequestID)
-	req, err := http.NewRequestWithContext(ctx, "GET", endpoint, nil)
-	if err != nil {
-		return nil, 0, err
-	}
-	bytes, statusCode, err := c.doRequest(req)
+	var dataRes PaymentRequestResponse
+	statusCode, err := c.doRequest(ctx, endpoint, "GET", &dataRes, nil)
 	if err != nil {
 		return nil, statusCode, err
-	}
-	var dataRes PaymentRequestResponse
-	err = json.Unmarshal(bytes, &dataRes)
-	if err != nil {
-		return nil, 0, err
 	}
 	return &dataRes, statusCode, nil
 }
 
 func (s *Store) GetPaymentRequest(ctx context.Context, paymentRequestID *PaymentRequestID) (*PaymentRequestResponse, int, error) {
 	endpoint := fmt.Sprintf("%s/api/v1/stores/%s/payment-requests/%s", s.Client.URL, s.ID, *paymentRequestID)
-	req, err := http.NewRequestWithContext(ctx, "GET", endpoint, nil)
-	if err != nil {
-		return nil, 0, err
-	}
-	bytes, statusCode, err := s.Client.doRequest(req)
+	var dataRes PaymentRequestResponse
+	statusCode, err := s.Client.doRequest(ctx, endpoint, "GET", &dataRes, nil)
 	if err != nil {
 		return nil, statusCode, err
-	}
-	var dataRes PaymentRequestResponse
-	err = json.Unmarshal(bytes, &dataRes)
-	if err != nil {
-		return nil, 0, err
 	}
 	return &dataRes, statusCode, nil
 }
 
 func (p *PaymentRequest) GetPaymentRequest(ctx context.Context) (*PaymentRequestResponse, int, error) {
 	endpoint := fmt.Sprintf("%s/api/v1/stores/%s/payment-requests/%s", p.Client.URL, p.Store.ID, p.ID)
-	req, err := http.NewRequestWithContext(ctx, "GET", endpoint, nil)
-	if err != nil {
-		return nil, 0, err
-	}
-	bytes, statusCode, err := p.Client.doRequest(req)
+	var dataRes PaymentRequestResponse
+	statusCode, err := p.Client.doRequest(ctx, endpoint, "GET", &dataRes, nil)
 	if err != nil {
 		return nil, statusCode, err
-	}
-	var dataRes PaymentRequestResponse
-	err = json.Unmarshal(bytes, &dataRes)
-	if err != nil {
-		return nil, 0, err
 	}
 	return &dataRes, statusCode, nil
 }
@@ -193,11 +81,7 @@ func (p *PaymentRequest) GetPaymentRequest(ctx context.Context) (*PaymentRequest
 // Archives the specified payment request.
 func (c *Client) ArchivePaymentRequest(ctx context.Context, storeID *StoreID, paymentRequestID *PaymentRequestID) (int, error) {
 	endpoint := fmt.Sprintf("%s/api/v1/stores/%s/payment-requests/%s", c.URL, *storeID, *paymentRequestID)
-	req, err := http.NewRequestWithContext(ctx, "DELETE", endpoint, nil)
-	if err != nil {
-		return 0, err
-	}
-	_, statusCode, err := c.doRequest(req)
+	statusCode, err := c.doRequest(ctx, endpoint, "DELETE", nil, nil)
 	if err != nil {
 		return statusCode, err
 	}
@@ -206,11 +90,7 @@ func (c *Client) ArchivePaymentRequest(ctx context.Context, storeID *StoreID, pa
 
 func (s *Store) ArchivePaymentRequest(ctx context.Context, paymentRequestID *PaymentRequestID) (int, error) {
 	endpoint := fmt.Sprintf("%s/api/v1/stores/%s/payment-requests/%s", s.Client.URL, s.ID, *paymentRequestID)
-	req, err := http.NewRequestWithContext(ctx, "DELETE", endpoint, nil)
-	if err != nil {
-		return 0, err
-	}
-	_, statusCode, err := s.Client.doRequest(req)
+	statusCode, err := s.Client.doRequest(ctx, endpoint, "DELETE", nil, nil)
 	if err != nil {
 		return statusCode, err
 	}
@@ -219,11 +99,7 @@ func (s *Store) ArchivePaymentRequest(ctx context.Context, paymentRequestID *Pay
 
 func (p *PaymentRequest) ArchivePaymentRequest(ctx context.Context) (int, error) {
 	endpoint := fmt.Sprintf("%s/api/v1/stores/%s/payment-requests/%s", p.Client.URL, p.Store.ID, p.ID)
-	req, err := http.NewRequestWithContext(ctx, "DELETE", endpoint, nil)
-	if err != nil {
-		return 0, err
-	}
-	_, statusCode, err := p.Client.doRequest(req)
+	statusCode, err := p.Client.doRequest(ctx, endpoint, "DELETE", nil, nil)
 	if err != nil {
 		return statusCode, err
 	}
@@ -233,66 +109,30 @@ func (p *PaymentRequest) ArchivePaymentRequest(ctx context.Context) (int, error)
 // Update a payment request
 func (c *Client) UpdatePaymentRequest(ctx context.Context, storeID *StoreID, paymentRequestID *PaymentRequestID, paymentRequestUpdate *PaymentRequestRequest) (*PaymentRequestResponse, int, error) {
 	endpoint := fmt.Sprintf("%s/api/v1/stores/%s/payment-requests/%s", c.URL, *storeID, *paymentRequestID)
-	dataReq, err := json.Marshal(paymentRequestUpdate)
-	if err != nil {
-		return nil, 0, err
-	}
-	req, err := http.NewRequestWithContext(ctx, "PUT", endpoint, bytes.NewBuffer(dataReq))
-	if err != nil {
-		return nil, 0, err
-	}
-	bytes, statusCode, err := c.doRequest(req)
+	var dataRes PaymentRequestResponse
+	statusCode, err := c.doRequest(ctx, endpoint, "PUT", &dataRes, paymentRequestUpdate)
 	if err != nil {
 		return nil, statusCode, err
-	}
-	var dataRes PaymentRequestResponse
-	err = json.Unmarshal(bytes, &dataRes)
-	if err != nil {
-		return nil, 0, err
 	}
 	return &dataRes, statusCode, nil
 }
 
 func (s *Store) UpdatePaymentRequest(ctx context.Context, paymentRequestID *PaymentRequestID, paymentRequestUpdate *PaymentRequestRequest) (*PaymentRequestResponse, int, error) {
 	endpoint := fmt.Sprintf("%s/api/v1/stores/%s/payment-requests/%s", s.Client.URL, s.ID, *paymentRequestID)
-	dataReq, err := json.Marshal(paymentRequestUpdate)
-	if err != nil {
-		return nil, 0, err
-	}
-	req, err := http.NewRequestWithContext(ctx, "PUT", endpoint, bytes.NewBuffer(dataReq))
-	if err != nil {
-		return nil, 0, err
-	}
-	bytes, statusCode, err := s.Client.doRequest(req)
+	var dataRes PaymentRequestResponse
+	statusCode, err := s.Client.doRequest(ctx, endpoint, "PUT", &dataRes, paymentRequestUpdate)
 	if err != nil {
 		return nil, statusCode, err
-	}
-	var dataRes PaymentRequestResponse
-	err = json.Unmarshal(bytes, &dataRes)
-	if err != nil {
-		return nil, 0, err
 	}
 	return &dataRes, statusCode, nil
 }
 
 func (p *PaymentRequest) UpdatePaymentRequest(ctx context.Context, paymentRequestUpdate *PaymentRequestRequest) (*PaymentRequestResponse, int, error) {
 	endpoint := fmt.Sprintf("%s/api/v1/stores/%s/payment-requests/%s", p.Client.URL, p.Store.ID, p.ID)
-	dataReq, err := json.Marshal(paymentRequestUpdate)
-	if err != nil {
-		return nil, 0, err
-	}
-	req, err := http.NewRequestWithContext(ctx, "PUT", endpoint, bytes.NewBuffer(dataReq))
-	if err != nil {
-		return nil, 0, err
-	}
-	bytes, statusCode, err := p.Client.doRequest(req)
+	var dataRes PaymentRequestResponse
+	statusCode, err := p.Client.doRequest(ctx, endpoint, "PUT", &dataRes, paymentRequestUpdate)
 	if err != nil {
 		return nil, statusCode, err
-	}
-	var dataRes PaymentRequestResponse
-	err = json.Unmarshal(bytes, &dataRes)
-	if err != nil {
-		return nil, 0, err
 	}
 	return &dataRes, statusCode, nil
 }
